Casts and scripts neatened

+++ utils/RCBot2_meta/bot_dod_mod.cpp
@@ -112,9 +112,9 @@ bool CDODMod :: shouldAttack ( int iTeam )
 	//TODO: Improve on the integer and floating point precision conversion [APG]RoboCop[CL]
 	iNumFlags = m_Flags.getNumFlags();
 
-	iFlags_0 = static_cast<int>(static_cast<float>(m_Flags.getNumFlagsOwned(iTeam == TEAM_ALLIES ? TEAM_AXIS : TEAM_ALLIES)) /
+	iFlags_0 = static_cast<short>(m_Flags.getNumFlagsOwned(iTeam == TEAM_ALLIES ? TEAM_AXIS : TEAM_ALLIES) /
 		iNumFlags * MAX_DOD_FLAGS);
-	iFlags_1 = static_cast<int>(static_cast<float>(m_Flags.getNumFlagsOwned(iTeam)) / iNumFlags * MAX_DOD_FLAGS);
+	iFlags_1 = static_cast<short>(m_Flags.getNumFlagsOwned(iTeam)) / (iNumFlags * MAX_DOD_FLAGS);
 
 	return randomFloat(0.0f,1.0f) < fAttackProbLookUp[iFlags_0][iFlags_1];//gNetAttackOrDefend->getOutput();
 }
@@ -160,8 +160,8 @@ void CDODMod :: initMod ()
 		{
 			tset.init();
 			tset.addSet();
-			tset.in(static_cast<float>(i) / MAX_DOD_FLAGS);
-			tset.in(static_cast<float>(j) / MAX_DOD_FLAGS);
+			tset.in(i / MAX_DOD_FLAGS);
+			tset.in(j / MAX_DOD_FLAGS);
 			nn.execute(tset.getBatches()->in,&fAttackProbLookUp[i][j],0.0f,1.0f);
 		}
 	}
@@ -494,12 +494,12 @@ bool CDODFlags:: getRandomBombToPlant ( CBot *pBot, Vector *position, int iTeam,
 
 	for ( short int i = 0; i < m_iNumControlPoints; i ++ )
 	{
-		if ( m_iWaypoint[i] != -1 )
+		if (m_iWaypoint[i] != -1)
 		{
-			if ( m_pBombs[i][0] == nullptr || m_iOwner[i] == iTeam || isBombPlanted(i) )
+			if (m_pBombs[i][0] == nullptr || m_iOwner[i] == iTeam || isBombPlanted(i))
 				continue;
 
-				fTotal += (MAX_BELIEF + 1.0f - pNav->getBelief(m_iWaypoint[i])) / MAX_BELIEF * getNumBombsRemaining(i);
+			fTotal += (MAX_BELIEF + 1.0f - pNav->getBelief(m_iWaypoint[i])) / MAX_BELIEF * getNumBombsRemaining(i);
 		}
 		else
 			fTotal += 0.1f;
@@ -627,7 +627,7 @@ int CDODFlags::setup(edict_t *pResourceEntity)
 
 		Vector vOrigin;
 
-		short int i = gpGlobals->maxClients;
+		int i = gpGlobals->maxClients;
 
 		// find visible flags -- with a model
 		while ( ++i < gpGlobals->maxEntities &&  m_pFlags[j] == nullptr)
@@ -911,7 +911,7 @@ void CDODMod ::clientCommand( edict_t *pEntity, int argc, const char *pcmd, cons
 	{
 		if ( std::strncmp(pcmd,"voice_",6) == 0 )
 		{
-			for ( short int i = 0; i < DOD_VC_INVALID; i ++ )
+			for ( byte i = 0; i < DOD_VC_INVALID; i ++ )
 			{
 				if ( std::strcmp(&pcmd[6],g_DODVoiceCommands[i].pcmd) == 0 )
 				{
